tools/requirements.txt
mysql-connector-python==8.1.0
flask==2.3.3